OBGM-47 Refactor ProductAssociationExcelImporter
